test(DateTimePicker): 增加单元测试
